requirements.txt
openai
python-pptx
python-dotenv
Pillow
requests
fastapi
uvicorn[standard]
python-multipart
aiofiles
pymupdf
pandas
